Move spill-refund tests into existing EIP-8037 suite files
